Build improvements

+++ update.sh
@@ -19,9 +19,13 @@ fail_check()
 
 pushd src
 
-fail_check wget -N https://publicsuffix.org/list/public_suffix_list.dat
-fail_check perl ../parse_rules.pl > publicsuffix_temp.erl
-fail_check mv publicsuffix_temp.erl publicsuffix.erl
-fail_check rm public_suffix_list.dat
+if [ -f "publicsuffix.erl" ]; then
+    echo "module already exist. remove publicsuffix.erl for a fresh generation ..."
+else
+    fail_check wget -N https://publicsuffix.org/list/public_suffix_list.dat
+    fail_check perl ../parse_rules.pl > publicsuffix_temp.erl
+    fail_check mv publicsuffix_temp.erl publicsuffix.erl
+    fail_check rm public_suffix_list.dat
+fi
 
 popd
